refactor(backend): move location of stuff
